Flowc. Fix debugger crash
Debugger crashes, because it is possible to mess global var
and function call during initialization:

globalVar : string = initGlobalVar();
initGlobalVar() { ""; }

Type checker takes this into account, but FcModule was not
modified and at generator stage it there was wrong order.

Normally QtByteRunner respects order of initialization
(call order), but debugger does it in a different way and fails.

This fix moves init order creation to desugar stage.