Allow to set flowc1 XMX via FLOWC_XMX env var
And expand set -e to set -euo pipefail.
set -e is to fail on any error
set -u is to fail on undefined variable
set -o pipefail prevents piped command return code masking. If any command in the
piped chain fails, the return code would be the return code of that command.